# swagger/level.yaml
openapi: 3.0.0
info:
  description: Monitoring Type API
  version: "1.0.0"
  title: Monitoring Type API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
  - name: admins
    description: Secured Admin-only calls
  - name: developers
    description: Operations available to regular developers
paths:
  /level:
    get:
      summary: Get all Levels
      operationId: getAlllevels
      description: get all existantes level in the system
      responses:
        "200":
          description: return all levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Level"
        "400":
          description: error message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

        "401":
          description: missing authentification
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: authentification required
    post:
      summary: adds a Level
      operationId: addLevel
      description: Adds Level to the system
      responses:
        "201":
          description: Level created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level"

        "400":
          description: error message
          content:
            application/json:
              schema:
                type: object
                properties:
                  messsage:
                    type: string
        "401":
          description: missing authentification
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: authentification required

      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - label
              properties:
                label:
                  type: number
                  example: 1
        description: Level item to add

  /level/{id}:
    patch:
      summary: update a single Level
      description: update one Level
      operationId: updateLevel
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string

      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level"

          description: "updated Level"
        "400":
          description: error message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: missing authentification
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: authentification required
        "404":
          description: Level not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Level not found
        "501":
          description: "no error displayed"

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: number
                  example: 1

components:
  schemas:
    Level:
      type: object
      required:
        - id
        - label
      properties:
        id:
          type: string
          example: "60edffc2370e9734982cc555"
        label:
          type: number
          example: 1

servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/skander12354/MonitoringActivity/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/skander12354/Type/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/skander12354/sysclass/1.0.0